+++ fe/fe-core/src/main/java/org/apache/doris/catalog/AggregateType.java
@@ -23,7 +23,9 @@ import com.google.common.collect.Lists;
 
 import java.util.EnumMap;
 import java.util.EnumSet;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 public enum AggregateType {
     SUM("SUM"),
@@ -38,6 +40,20 @@ public enum AggregateType {
     GENERIC_AGGREGATION("GENERIC_AGGREGATION");
 
     private static EnumMap<AggregateType, EnumSet<PrimitiveType>> compatibilityMap;
+
+    private static final Map<String, AggregateType> aggTypeMap = new HashMap<>();
+
+    static {
+        aggTypeMap.put("NONE", AggregateType.NONE);
+        aggTypeMap.put("SUM", AggregateType.SUM);
+        aggTypeMap.put("MIN", AggregateType.MIN);
+        aggTypeMap.put("MAX", AggregateType.MAX);
+        aggTypeMap.put("REPLACE", AggregateType.REPLACE);
+        aggTypeMap.put("REPLACE_IF_NOT_NULL", AggregateType.REPLACE_IF_NOT_NULL);
+        aggTypeMap.put("HLL_UNION", AggregateType.HLL_UNION);
+        aggTypeMap.put("BITMAP_UNION", AggregateType.BITMAP_UNION);
+        aggTypeMap.put("QUANTILE_UNION", AggregateType.QUANTILE_UNION);
+    }
 
     static {
         compatibilityMap = new EnumMap<>(AggregateType.class);
@@ -180,5 +196,9 @@ public enum AggregateType {
             default:
                 return null;
         }
+    }
+
+    public static AggregateType getAggTypeFromAggName(String typeName) {
+        return aggTypeMap.get(typeName);
     }
 }
