+++ fe/fe-core/src/main/java/org/apache/doris/common/proc/RemoteIndexSchemaProcDir.java
@@ -71,7 +71,7 @@ public class RemoteIndexSchemaProcDir implements ProcDirInterface {
         if (remoteSchema == null || remoteSchema.isEmpty()) {
             throw new AnalysisException("fetch remote tablet schema failed");
         }
-        this.schema = remoteSchema;
+        this.schema.addAll(remoteSchema);
         return IndexSchemaProcNode.createResult(this.schema, this.bfColumns);
     }
